Reviewed implicit cast warnings to determine safe casting and added explicit casts where safe or updated receiving variable to match datatype.
git-svn-id: https://crosswire.org/svn/sword/trunk@3439 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ include/listkey.h
@@ -144,7 +144,7 @@ public:
 	 * Returns the index for the new one given as as parameter.
 	 * The first parameter is the new index.
 	 */
-	virtual void setIndex(long index) { setToElement(index); }
+	virtual void setIndex(long index) { setToElement((int)index); }
 	virtual const char *getText() const;
 	virtual void setText(const char *ikey);
 	virtual void sort();
